Spectroelectrochemistry for Opentrons OT-2 Robot
STL file of the section of the robot that does the spectroelectrochemistry measurements. There are two sections that the cuvette can be placed in, depending on whether you are performing absorption or PL measurements. Those are in the upper right-hand corner. The bottom left corner is where the laser is intended to go, with a spot to insert a filter if needed. To the right of the laser holder, a Thor Labs filter cube is designed to be placed, directing the excitation source to the sample and then the emissions to the spectrometer.